ISA test for UART0.
Added a new test to isa/soc that exercises UART0 registers and sending
data in a loopback. The loopback is create at the testbench level by
connecting TxD to RxD.

As it is expected that more peripheral tests will follow, common defines
were pushed out to a separate include file (soc.h).